backend/requirements.txt
aiodns==3.2.0
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
blinker==1.7.0
boto3==1.34.84
botocore==1.34.84
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.1
Flask==3.0.3
Flask-Cors==4.0.0
frozenlist==1.4.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
inquirerpy==0.3.4
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
multidict==6.0.5
orjson==3.10.1
packaging==24.0
paramiko==3.4.0
pfzy==0.3.4
pluggy==1.4.0
prettytable==3.10.0
prompt-toolkit==3.0.43
py-cpuinfo==9.0.0
pycares==4.4.0
pycparser==2.22
pydantic==2.7.0
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
PyNaCl==1.5.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
runpod==1.6.2
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tomli==2.0.1
tomlkit==0.12.4
tqdm==4.66.2
tqdm-loggable==0.2
typing_extensions==4.11.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.2
yarl==1.9.4
